Consolidate coroutine config macros
